+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 2.9.11
 - Fix PHP fatal error: check Include file exists in StudentFields.php & UserFields.php
 - Fix SQL error unterminated quoted string at or near "'1 in EditReportCardGrades.php
 - Added common file types to upload white list in PortalNotes.php & StudentAssignments.fnc.php
+- Help texts updates in Help_en.php & Help_es.php
 
 Changes in 2.9.10
 -----------------